Add the ability to switch PyQt version
PyQt4 still the default, PyQt5 needs run build for updating the ui files
Also need set QT_BINDING=PyQt5 env to run with PyQt5

+++ setup.py
@@ -9,8 +9,12 @@ from distutils.spawn import find_executable, spawn
 from distutils.errors import DistutilsExecError
 from glob import glob
 
-from PyQt4 import uic
+from Qt import uic
+from Qt import QtCore
 from version import VERSION
+
+
+IS_QT5 = QtCore.QT_VERSION >= 0x050000
 
 
 class CustomBuild(build):
@@ -44,7 +48,8 @@ class BuildQt(Command):
             pyFile.close()
 
     def compile_ts(self):
-        lrelease = find_executable("lrelease-qt4") or \
+        lrelease_bin = "lrelease-qt5" if IS_QT5 else "lrelease-qt4"
+        lrelease = find_executable(lrelease_bin) or \
             find_executable("lrelease")
         if not lrelease:
             raise DistutilsExecError("Missing lrelease")
@@ -64,9 +69,10 @@ class UpdateTs(Command):
         pass
 
     def run(self):
-        lupdate = find_executable("pylupdate4")
+        lupdate_bin = "pylupdate5" if IS_QT5 else "pylupdate4"
+        lupdate = find_executable(lupdate_bin)
         if not lupdate:
-            raise DistutilsExecError("Missing pylupdate4")
+            raise DistutilsExecError("Missing %s" % lupdate_bin)
 
         path = os.path.realpath("data/translations/gitc.pro")
         spawn([lupdate, path])
@@ -92,13 +98,14 @@ class BuildExe(Command):
             dirName = "exe.%s-%s" % (distutils.util.get_platform(),
                                      sys.version[0:3])
             baseDir = "build\\" + dirName
+            qtDir = "PyQt5" if IS_QT5 else "PyQt4"
             files = [baseDir + "\\Qt*.dll",
-                     baseDir + "\\PyQt4\\*.exe"]
+                     baseDir + "\\%s\\*.exe" % qtDir]
             for pattern in files:
                 for file in glob(pattern):
                     os.remove(file)
 
-            baseDir += "\\PyQt4"
+            baseDir += "\\" + qtDir
             dirs = ["doc", "examples", "include", "mkspecs",
                     "uic"]
 
@@ -107,19 +114,24 @@ class BuildExe(Command):
                 if os.path.exists(fullPath):
                     shutil.rmtree(fullPath)
 
-            files = ["QtWebKit4.dll", "QtDesigner4.dll",
-                     "QtDeclarative4.dll", "QtDesignerComponents4.dll",
-                     "QtScript4.dll", "QtCLucene4.dll",
-                     "QtScriptTools4.dll", "QtHelp4.dll",
-                     "QtTest4.dll"]
+            if IS_QT5:
+                # TODO:
+                files = []
+            else:
+                files = ["QtWebKit4.dll", "QtDesigner4.dll",
+                         "QtDeclarative4.dll", "QtDesignerComponents4.dll",
+                         "QtScript4.dll", "QtCLucene4.dll",
+                         "QtScriptTools4.dll", "QtHelp4.dll",
+                         "QtTest4.dll"]
             for file in files:
                 fullPath = baseDir + "\\" + file
                 if os.path.exists(fullPath):
                     os.remove(fullPath)
 
-                fullPath = fullPath.replace("4.dll", ".pyd")
-                if os.path.exists(fullPath):
-                    os.remove(fullPath)
+                if not IS_QT5:
+                    fullPath = fullPath.replace("4.dll", ".pyd")
+                    if os.path.exists(fullPath):
+                        os.remove(fullPath)
 
 
 setup(name='gitc',
